refactor: simplify release plan diff helpers
