+++ pkg/microservice/aslan/core/release_plan/service/openapi.go
@@ -387,11 +387,11 @@ func OpenAPIUpdateReleasePlanWithJobs(c *handler.Context, id string, rawArgs *Op
 	if err != nil {
 		return errors.Wrap(err, "build release plan current snapshot")
 	}
-	hasChanges, err := hasReleasePlanPersistedSectionChanges(originalPlan, releasePlanVersionSectionPlan, currentSnapshot)
+	persistedSnapshot, err := buildReleasePlanVersionSnapshot(originalPlan, releasePlanVersionSectionPlan)
 	if err != nil {
 		return errors.Wrap(err, "build release plan persisted snapshot")
 	}
-	if !hasChanges {
+	if !hasReleasePlanSnapshotChanges(persistedSnapshot, currentSnapshot) {
 		return nil
 	}
 	var baseSnapshot interface{}
@@ -400,37 +400,27 @@ func OpenAPIUpdateReleasePlanWithJobs(c *handler.Context, id string, rawArgs *Op
 		return errors.Wrap(err, "check release plan base snapshot")
 	}
 	if needBaseSnapshot {
-		baseSnapshot, err = buildReleasePlanInputSnapshot(originalPlan)
-		if err != nil {
-			return errors.Wrap(err, "build release plan base snapshot")
-		}
+		baseSnapshot = persistedSnapshot
 	}
-	logBaseSnapshot, err := resolveReleasePlanLogBaseSnapshot(baseSnapshot, originalPlan, releasePlanVersionSectionPlan)
-	if err != nil {
-		return errors.Wrap(err, "build release plan log base snapshot")
-	}
-	shouldCreateLog := hasReleasePlanSnapshotChanges(logBaseSnapshot, currentSnapshot)
 
 	versionDoc := newReleasePlanVersionDocument(plan.ID.Hex(), plan.Version, previousVersion, baseSnapshot, currentSnapshot, c.UserName, c.Account, releasePlanVersionSectionPlan, releasePlanVersionSectionName(releasePlanVersionSectionPlan, plan.Name), string(VerbUpdate))
 	if err := persistReleasePlanWithVersion(c, id, plan, versionDoc); err != nil {
 		return errors.Wrap(err, "persist release plan update")
 	}
-	if shouldCreateLog {
-		if err := createReleasePlanLog(&models.ReleasePlanLog{
-			PlanID:      plan.ID.Hex(),
-			Username:    c.UserName,
-			Account:     c.Account,
-			Verb:        VerbUpdate,
-			TargetName:  plan.Name,
-			TargetType:  TargetTypeReleasePlan,
-			Version:     plan.Version,
-			SectionKey:  releasePlanVersionSectionPlan,
-			SectionName: releasePlanVersionSectionName(releasePlanVersionSectionPlan, plan.Name),
-			SectionType: releasePlanVersionSectionGroupType(releasePlanVersionSectionPlan),
-			CreatedAt:   time.Now().Unix(),
-		}); err != nil {
-			log.Errorf("create release plan log error: %v", err)
-		}
+	if err := createReleasePlanLog(&models.ReleasePlanLog{
+		PlanID:      plan.ID.Hex(),
+		Username:    c.UserName,
+		Account:     c.Account,
+		Verb:        VerbUpdate,
+		TargetName:  plan.Name,
+		TargetType:  TargetTypeReleasePlan,
+		Version:     plan.Version,
+		SectionKey:  releasePlanVersionSectionPlan,
+		SectionName: releasePlanVersionSectionName(releasePlanVersionSectionPlan, plan.Name),
+		SectionType: releasePlanVersionSectionGroupType(releasePlanVersionSectionPlan),
+		CreatedAt:   time.Now().Unix(),
+	}); err != nil {
+		log.Errorf("create release plan log error: %v", err)
 	}
 	if err := broadcastReleasePlanCollaboration(plan.ID.Hex()); err != nil {
 		log.Errorf("broadcast release plan collaboration error: %v", err)
